kittygram/urls.py
from rest_framework.routers import DefaultRouter
from django.contrib import admin
from django.urls import include, path

from cats.views import CatViewSet, LightCatViewSet, OwnerViewSet

router = DefaultRouter()
router.register('cats', CatViewSet)
router.register('owners', OwnerViewSet)
router.register(r'mycats', LightCatViewSet)

urlpatterns = [
    path('auth/', include('djoser.urls')),
    path('auth/', include('djoser.urls.jwt')),
    path('', include(router.urls), name='api-root'),
    path('admin/', admin.site.urls),
]
